backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-socketio==5.11.0
google-generativeai==0.7.2
pydantic==2.9.0
pydantic-settings==2.5.0
thefuzz==0.22.1
python-multipart==0.0.9
httpx==0.27.2
Pillow==10.4.0
aiofiles==24.1.0
yt-dlp==2024.11.04
redis[hiredis]==5.2.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
asyncpg==0.30.0
pytest==8.3.4
pytest-asyncio==0.25.2
